Address review: Move upsel UI to `admin/upsells`
