--- frontend/src/metabase/public/components/EmbedModal/StaticEmbedSetupPane/LookAndFeelSettings/index.ts
@@ -1,1 +1,0 @@
-export { LookAndFeelSettings } from "./LookAndFeelSettings";
